==> notebooks/data_prep/01_data_loading.ipynb <==
import sys
sys.path.append('../../')

from src.load_datasets  import load_csv_files

## Visualisation des Datasets

# Exemple d'utilisation de la fonction
dfs = load_csv_files("../../data/inputs")

print(dfs['circuits'].shape)
dfs['circuits'].head()

print(dfs['lap_times'].shape)
dfs['lap_times'].head()

print(dfs['drivers'].shape)
dfs['drivers'].head()

print(dfs['races'].shape)
dfs['races'].head()


print(dfs['constructors'].shape)
dfs['constructors'].head()

dfs['constructors'][dfs['constructors'].name=='Alpine']


dfs['constructor_standings'].head()


print(dfs['driver_standings'].shape)
dfs['driver_standings'].head()


print(dfs['constructor_results'].shape)
dfs['constructor_results'].tail(10)

print(dfs['seasons'].shape)
dfs['seasons'].head()


print(dfs['results'].shape)
dfs['results'].head()
