==> notebooks/data_prep/02_result_processing.ipynb <==
import sys
sys.path.append('../../')

import pandas as pd

from src.time_transformer import time_to_milliseconds
from src.transformer_utils import fill_na, print_df_missing_values, set_the_right_col_type

## 1. Chargement du Dataset

df_results = pd.read_csv('../../data/inputs/results.csv')

print(df_results.shape)
df_results.head()

df_results.info()

## 2. Traitement des valeurs nulles
### 2.1 Recherchons les valeures nulles
En se référent à la documentation du jeu de donnée, nous remarquons que les valeures nulles sont plutot représentés par la valeur \N

car = '\\N'
print_df_missing_values(df_results, car)

df_results.isnull().sum()

### 2.2. Ajouter une valeur par défaut dans les lignes qui ont la valeur \N

cols = ['milliseconds', 'fastestLap', 'fastestLapSpeed']
# Etant donnée que ces trois champs sont quasiment du meme types nous pourrons leur attribuer la même valeur par défaut
for field in cols:
    df_results = fill_na(df_results, field, 0)

# Ajoutons la valeur par defaut à 90 min
df_results = fill_na(df_results, 'fastestLapTime', '90:00')

car = '\\N'
print_df_missing_values(df_results, car)

## 2. Convertir le le champs fastLapTime en millisecondes

### 2.2. Convert a fastestLapTime field

# Convertion de 'fastestLapTime' en millisecondes en utilisant une opération vectorielle
df_results['fastestLapTime'] = df_results['fastestLapTime'].apply(time_to_milliseconds)

df_results.head()

## 3. Retirer les champs unitilisés

Nous pourrons avoir ces informations dans l'ordre de position. Cette information est présente dans le dataset des pilotes

df_results.drop(['positionText', 'number', 'position', 'rank', 'time'], axis=1, inplace=True)

df_results.head()

## 4. Int fields to int64

df_results = set_the_right_col_type(df_results, ['milliseconds', 'fastestLap'], 'int64')
df_results = set_the_right_col_type(df_results, ['fastestLapSpeed'], 'float64')

df_results.info()

df_results.to_csv('../../data/outputs/01_results_cleaned.csv', index=False)
