==> notebooks/data_prep/03_circuit_processing.ipynb <==
import sys
sys.path.append('../../')

import pandas as pd

## 1. Loading the dataset

df_circuits = pd.read_csv('../../data/inputs/circuits.csv')

print(df_circuits.shape)
df_circuits.head()

## 2. Getting the location list for all the circuite

location_list = ('Sakhir','Jeddah','Melbourne','Suzuka','Shanghai','Miami','Imola', 'Monte-Carlo', 'Montreal', 'Montmeló',
                 'Spielberg', 'Silverstone', 'Budapest', 'Spa', 'Zandvoort', 'Monza', 'Baku', 'Marina Bay', 'Austin',
                 'Mexico City', 'São Paulo', 'Las Vegas', 'Al Daayen', 'Abu Dhabi')

## 3. Adding the active flag in the current actives circuits

df_circuits['circuits_is_active'] = 0

for loc in location_list:
    df_circuits.loc[df_circuits['location'].str.contains(loc, case=False), 'circuits_is_active'] = 1

print(df_circuits[df_circuits.circuits_is_active == 1].shape)
df_circuits[df_circuits.circuits_is_active == 1]

## 4. Export the dataset

df_circuits.to_csv('../../data/outputs/02_circuits_processed.csv', index=False)
