==> notebooks/data_prep/03_merging_race_and_laps.ipynb <==
import pandas as pd

## 1. Chargement du Dataset

df_races = pd.read_csv('../../data/inputs/races.csv')
df_laps = pd.read_csv('../../data/inputs/lap_times.csv')

print(df_races.shape)
df_races.head()

print(df_laps.shape)
df_laps.head()

## 2. Renomons la colonne positionOrder par race_rank 

df_laps.rename({'position': 'lap_position'}, axis=1, inplace=True)
df_laps.rename({'milliseconds': 'lap_time'}, axis=1, inplace=True)

df_laps.head()

## 2. Associons les deux jeux de données

df_race_laps = df_laps.merge(df_races[['raceId', 'year', 'round', 'circuitId', 'name']], how='left', on='raceId')

print(df_race_laps.shape)
df_race_laps.head()

## 3. Suppression de colonne non utilisés
Nous allons supprimer les colonnes
- time parce que cette information est deja présente dans la colonne lap_time 

df_race_laps.drop(['time'], axis=1, inplace=True)

df_race_laps.head()

## 4. Exporter le Dataset

df_race_laps.to_csv("../../data/outputs/02_races_laps.csv", index=False)
