==> notebooks/data_prep/04_merging_results_and_race.ipynb <==
import pandas as pd

## 1. Chargement du Dataset

df_races = pd.read_csv('../../data/inputs/races.csv')
df_results = pd.read_csv('../../data/outputs/01_results_cleaned.csv')

print(df_races.shape)
df_races.head()

print(df_results.shape)
df_results.head()

## 2. Renomons la colonne positionOrder par race_rank 

df_results.rename({'positionOrder': 'race_rank'}, axis=1, inplace=True)

## 3. Associons les deux jeux de données

df_race_results = df_results.merge(df_races[['raceId', 'year', 'round', 'circuitId', 'name']], how='inner', on='raceId')

print(df_race_results.shape)
df_race_results.head()

## 4. Exporter le Dataset

df_race_results.to_csv("../../data/outputs/03_races_results.csv", index=False)
