==> notebooks/data_prep/05_driver_processing.ipynb <==
Traitements à faire
---
- Retirer les colonnes inutilisées
- Ajouter une étique pour les pilotes actifs à la dernière saison (2023)

import pandas as pd

## 1. Chargement du Dataset

df_drivers = pd.read_csv("../../data/inputs/drivers.csv")

df_drivers.head()

## 2. Retirer les colonnes inutilisées

df_drivers.drop(['driverRef', 'url', 'code'], axis=1, inplace=True)
df_drivers.head()

## 3. Ajouter une étique uniquement sur les pilotes encore actifs

df_drivers['driver_is_active'] = 0

# Liste des pilotes actifs actuels
list_drivers = [
    ("Max", "Verstappen"),
    ("Logan", "Sargeant"),
    ("Lando", "Norris"),
    ("Pierre", "Gasly"),
    ("Sergio", "Pérez"),
    ("Fernando", "Alonso"),
    ("Charles", "Leclerc"),
    ("Lance", "Stroll"),
    ("Kevin", "Magnussen"),
    ("Yuki", "Tsunoda"),
    ("Alexander", "Albon"),
    ("Esteban", "Ocon"),
    ("Lewis", "Hamilton"),
    ("Carlos", "Sainz"),
    ("George", "Russell"),
    ("Oscar", "Piastri"),
    ("Guanyu", "Zhou"),
    ("Nico", "Hülkenberg"),
    ("Valtteri", "Bottas"),
    ("Daniel", "Ricciardo"),
]

# Ajouter l'étique en fonction de la correspondance nom et prénom du pilote dans le Dataset
for (name, surname) in list_drivers:
    df_drivers.loc[(df_drivers['forename'] == name) & (df_drivers['surname'] == surname), 'driver_is_active'] = 1

df_drivers[df_drivers.driver_is_active == 1]

## 4. Exportons le Dataset

df_drivers.to_csv('../../data/outputs/03_drivers_processed.csv', index=False)
