==> notebooks/data_prep/06_constructor_processing.ipynb <==
Traitements à faire
---
- Retirer les colonnes inutilisées
- Ajouter une étique pour les constructeurs actifs à la dernière saison (2023)

import pandas as pd

## 1. Chargement du Dataset

df_constructors = pd.read_csv("../../data/inputs/constructors.csv")
df_races_results = pd.read_csv("../../data/outputs/03_races_results.csv")

print(df_constructors.shape)
df_constructors.head()

df_races_results.head()

## 2. Suppréssion des colonnes inutilisées

df_constructors.drop(['constructorRef', 'url'], axis=1, inplace=True)

df_constructors.head()

## 3. Etiquetage des constructeurs actifs

# Créer un jeu de donnée de transition qui permettra d'avoir la date sur le dataset de constructeur afin de mieux filtrer plus tard
df_transition = df_constructors.merge(df_races_results[['constructorId', 'year']], how='inner', on='constructorId')

df_transition.head()

active_constructors = df_transition[df_transition.year == 2023]['name'].unique().tolist()
print(active_constructors)

df_constructors['constructor_is_active'] = 0

for constructor in active_constructors:
    df_constructors.loc[df_constructors.name == constructor, 'constructor_is_active'] = 1

df_constructors[df_constructors.constructor_is_active == 1]

## 4. Exporter le Dataset

df_constructors.to_csv('../../data/outputs/04_constructors_processed.csv', index=False)
