==> notebooks/data_prep/07_merge_all_the_df.ipynb <==
Traitements à faire
---
- Fusion de tous les Datasets
- Fusion des colonnes forename et surname en une colonne 
- Calculer l'age du pilote à la date de la course
- Retirer les colonnes inutilisées
- Exporter le Dataset fusionné

import sys
sys.path.append('../../')

import pandas as pd

from src.transformer_utils import print_df_missing_values

## 1. Chargement des Datasets

df_constructors = pd.read_csv('../../data/outputs/04_constructors_processed.csv')
df_drivers = pd.read_csv('../../data/outputs/03_drivers_processed.csv')
df_races_results = pd.read_csv('../../data/outputs/03_races_results.csv')
df_circuits = pd.read_csv('../../data/outputs/02_circuits_processed.csv')

df_constructors.head()

df_drivers.head()

df_races_results.head()

df_circuits.head()

## 2. Fusion des Datasets

### 2.1. Renommons les colonnes dans les Datasets

Etant donné que nous avons une colonne name dans le Dataset des resultats et des constructeurs, il est necessaire pour nous de renommer ces colonnes dans chacun des Dataset

df_races_results.rename({'name': 'circuit_name'}, axis=1, inplace=True)
df_constructors.rename({'name': 'constructor_name'}, axis=1, inplace=True)
df_constructors.rename({'nationality': 'country'}, axis=1, inplace=True)

### 2.2. Fusion des Datasets constructeurs et résultats

df_const_race = df_races_results.merge(df_constructors, on='constructorId', how='inner')

df_const_race.head()

### 2.3. Fusion du Dataset des pilotes et des résultats des courses

# Visualisation du Dataset des pilotes
df_drivers.head()

#### a) Récupérons uniquement les colonnes utiles du Dataset des pilotes

cols = [x for x in df_drivers.columns if x not in ['number', 'code', 'nationality', 'url']]

df_drivers[cols].head()

#### b) Créons une colonne qui contient le nom et le prénom du pilote 

df_drivers['full_name'] = df_drivers['forename'] + ' ' + df_drivers['surname']
# Suppression des colonnes cars elles ne nous sont plus utiles
df_drivers.drop(['forename', 'surname', 'number'], axis=1, inplace=True)

#### c) Fusionnons tous les Datasets

df_merged = df_const_race.merge(df_drivers, on='driverId', how='inner')

print(df_merged.shape)
df_merged.head()

### 2.4. Fusion avec les circuits

df_merged = df_merged.merge(df_circuits[['circuitId', 'circuits_is_active']], on='circuitId', how='inner')

print(df_merged.shape)
df_merged.head()

## 4. Calcul de l'age de chaque pilote par course

def calcul_age(row: pd.DataFrame) -> int:
    return row['year'] - row['dob'].year

# Convertion de la colonne dob en type date Pandas
df_merged['dob'] = pd.to_datetime(df_merged['dob'])
df_merged['season_age'] = df_merged.apply(calcul_age, axis=1)

# Suppression de la colonne dob car elle ne sera plus utile pour la suite
df_merged.drop('dob', axis=1, inplace=True)

df_merged.head()

## 5. Verifions si on a aucune valeur manquante avant de poursuivre

car = '\\N'
print_df_missing_values(df_merged, car)

## 5. Exportons le Dataset

df_merged.to_csv('../../data/outputs/05_merged_data.csv', index=False)
